requirements.txt
Flask==2.3.3
Flask-Login==0.6.2
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.1.1
Werkzeug==2.3.7
SQLAlchemy==2.0.20
python-dotenv==1.0.0
WTForms==3.0.1
pdfminer.six==20221105
python-docx==0.8.11
spacy==3.6.1
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.6.0/en_core_web_sm-3.6.0-py3-none-any.whl
PyPDF2==3.0.1
Pillow==10.0.0
openpyxl==3.1.2